backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-cloud-firestore==2.13.1
gunicorn==21.2.0
